src/eva-wraith/pyproject.toml
[project]
name = "eva-wraith"
version = "0.1.0"
description = "The Wraith - Agent Vision & Perception THE HIVE"
requires-python = ">=3.11"
dependencies = [
    "shared",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["eva_wraith*"]
